+++ src/actions/track.action.ts
@@ -1,6 +1,13 @@
+import instanceLocker from "instance-locker";
 import { Tracker } from "../classes/tracker.class";
+const locker = instanceLocker("accel-shooter track");
 
 export async function trackAction() {
-  const tracker = new Tracker();
-  tracker.startSync();
+  const success = await locker.Lock();
+  if (success) {
+    const tracker = new Tracker();
+    tracker.startSync();
+  } else {
+    console.log("Lock occupied!");
+  }
 }